+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## [0.0.3] - 14 Nov 2019
+
+- introduction of offset and limit control variable in range search
+- custom comparator during search
+- range delete
+
 ## [0.0.2] - 4 Nov 2019
 * Included support for range search